Quarter final selection and scouting

+++ script/convert-schedules.pl
@@ -3,11 +3,11 @@
 use strict;
 use Data::Dumper;
 
-my $files = [ split(/\n/, `find . -name '*.dat'`) ];
+my $datFiles = [ split(/\n/, `find . -name '*.dat'`) ];
 
 sub pairsToCsv(){
     my ($content, $file) = @_;
-    my $csv = [["match","R1","R2","R2","B1","B2","B3"]];
+    my $csv = [["Match","R1","R2","R3","B1","B2","B3"]];
     for my $line (split(/\n/, $content)){
         if ($line =~ /_qm([0-9]+)_([1-6]) *= *([0-9]+)/){
             my ($match, $pos, $team) = (int($1),int($2),$3);
@@ -24,14 +24,45 @@ sub pairsToCsv(){
     return $content;
 }
 
-for my $file (@$files){
+for my $file (@$datFiles){
     my $content = `cat $file`;
     $content = &pairsToCsv($content, $file) if ($content =~ /\=/);
     my $newfile = $file;
     $newfile =~ s/.*\//www\/data\//g;
-    $newfile =~ s/\.dat$/.quals.csv/g;
-    `rm $file`;
+    $newfile =~ s/\.dat$/.schedule.csv/g;
     die("Error opening $file for writing: $!") if (!open my $fh, ">", $newfile);
     print $fh $content;
     close $fh;
+
+    `rm $file`;
+}
+
+my $elimsFiles = [ split(/\n/, `find . -name '*.elims'`) ];
+for my $file (@$elimsFiles){
+    my $content = `cat $file`;
+    $content =~ s/-/,/g;
+    my $alliances = [ split(/\n/, $content) ];
+    my $allianceFile = $file;
+    $allianceFile =~ s/.*\//www\/data\//g;
+    $allianceFile =~ s/\.elims$/.alliances.csv/g;
+    die("Error opening $file for writing: $!") if (!open my $fh, ">", $allianceFile);
+    print $fh "Alliance,Captain,First Pick,Second Pick\n";
+    for my $i (1..8){
+        print $fh "$i,".$alliances->[$i-1]."\n"
+    }
+    close $fh;
+
+    my $scheduleFile = $file;
+    $scheduleFile =~ s/.*\//www\/data\//g;
+    $scheduleFile =~ s/\.elims$/.schedule.csv/g;
+    die("Error opening $file for writing: $!") if (!open my $fh, ">>", $scheduleFile);
+    my $matches = [[1,8],[4,5],[2,7],[3,6],[1,8],[4,5],[2,7],[3,6],[1,8],[4,5],[2,7],[3,6]];
+    for my $i (1..scalar(@$matches)){
+        my $r = $alliances->[$matches->[$i-1]->[0]-1];
+        my $b = $alliances->[$matches->[$i-1]->[1]-1];
+        print $fh "qf$i,$r,$b\n"
+    }
+    close $fh;
+
+    `rm $file`;
 }
